KS-1-4 used Byte instead of Short

+++ src/main/kotlin/lesson1/lesson1_task4.kt
@@ -1,7 +1,7 @@
 package lesson1
 fun main(){
     val length: Long = 40868600000
-    val age: Short = 27
+    val age: Byte = 27
     val partOfDay: Float = 0.075F
     val seconds: Short = 6480
     val partOfYear: Double = 2.0547945205479453E-4
